Added a JavaScript function for algorithms.
Still needs CSS, but the JavaScript was the goal for this commit.
JSON string at the top for algorithm data. Not every algorithm is there yet. Feel free to contribute if you want to.
Call this function using (as an example) "getIntermediateAlgorithms('OLL')". This will get all intermediate oll algorithms and put them into the element with id "oll". This was meant to be used flexibly.
I updated the advanced page to reflect these changes as an example of how to implement this.