Added new algorithms and overhauled note system
Credit to w3schools.com for the css for tooltips! Added lots of algorithms. Images for algorithms are now in dedicated folders to make them easier to find. Old image system replaced by customPath. If customPath is commented (which is the default), the image's path will be images/algorithms/[set name]/[algorithm name]. Custom path lets you change this (currently, no algorithms use this, but it was quick to implement and I could see it being useful in the future). Added note system (by default note is blank or commented). When hovering your mouse over the notation for an algorithm (e.g. "F R U R' U' F'"), a note will pop up with help text. Credit to w3schools.com for the css. The current help text is mostly placeholder. There is a bug I'm not sure how to fix: if the help text is there, the notation takes the minimum amount of space it needs, not the maximum like it should. Advanced OLL algorithm 25 or Advanced PLL algorithm T (among others) are examples of this.