Algorithm notation is now an array
Before I say what I changed, I want to say: when solving a Rubik's Cube, there is usually more than one algorithm that can accomplish the same thing. One algorithm might be easier to learn than another for a specific person, so top solvers will often try multiple algorithms to see which works best for them.

Added a new function: switchAlgorithm(). This will switch which algorithm is displayed when you click on the notation for one. Currently, the only algorithm with multiple algorithms to choose from is OLL #2. Effectively, this does nothing to algorithms other than OLL #2, for now. I will make a more comprehensive list soon (the various algorithms are on the internet, but I need to copy/paste them over to the JSON string). It's just a matter of spending time to copy/paste.

Small change: the defaultAlgorithm attribute of each algorithm picks which one to choose on page load. By default it is zero. Does nothing when there is only one possible algorithm to choose from.