Added a function to say you learned an algorithm
When you click on the name of an algorithm (number for oll, 1-2 letters for pll) it will toggle being green. This is mostly prep for backend stuff.
The chartreuse color is placeholder right now, although I currently have no plans for changing it. If anyone is good with colors, feel free to change it (on lines 659 and 662).